BJ-691: add tests

+++ src/main/java/com/rbkmoney/dark/api/converter/konturfocus/ShareHolderFL.java
@@ -5,6 +5,8 @@ import com.rbkmoney.dark.api.converter.SwagConverterContext;
 import com.rbkmoney.swag.questionary_aggr_proxy.model.ShareHolderFl;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class ShareHolderFL implements SwagConverter<ShareHolderFl, com.rbkmoney.questionary_proxy_aggr.kontur_focus_egr_details.ShareHolderFl> {
 
@@ -12,10 +14,9 @@ public class ShareHolderFL implements SwagConverter<ShareHolderFl, com.rbkmoney.
     public ShareHolderFl toSwag(com.rbkmoney.questionary_proxy_aggr.kontur_focus_egr_details.ShareHolderFl value, SwagConverterContext ctx) {
         return new ShareHolderFl()
                 .address(value.getAddress())
-                .capitalSharesPercent((int) value.getCapitalSharesPercent())
+                .capitalSharesPercent(BigDecimal.valueOf(value.getCapitalSharesPercent()))
                 .date(value.getDate())
                 .fio(value.getFio())
-                .votingSharesPercent((int) value.getVotingSharesPercent());
+                .votingSharesPercent(BigDecimal.valueOf(value.getVotingSharesPercent()));
     }
-
 }
