+++ src/main/java/com/rbkmoney/dark/api/converter/konturfocus/ShareHolderOtherSwagConverter.java
@@ -5,6 +5,8 @@ import com.rbkmoney.dark.api.converter.SwagConverterContext;
 import com.rbkmoney.swag.questionary_aggr_proxy.model.ShareHolderOther;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class ShareHolderOtherSwagConverter implements SwagConverter<ShareHolderOther, com.rbkmoney.questionary_proxy_aggr.kontur_focus_egr_details.ShareHolderOther> {
 
@@ -12,10 +14,9 @@ public class ShareHolderOtherSwagConverter implements SwagConverter<ShareHolderO
     public ShareHolderOther toSwag(com.rbkmoney.questionary_proxy_aggr.kontur_focus_egr_details.ShareHolderOther value, SwagConverterContext ctx) {
         return new ShareHolderOther()
                 .address(value.getAddress())
-                .capitalSharesPercent((int) value.getCapitalSharesPercent())
+                .capitalSharesPercent(BigDecimal.valueOf(value.getCapitalSharesPercent()))
                 .date(value.getDate())
                 .fullName(value.getFullName())
-                .votingSharesPercent((int) value.getVotingSharesPercent());
+                .votingSharesPercent(BigDecimal.valueOf(value.getVotingSharesPercent()));
     }
-
 }
